@@ -6,21 +6,22 @@ import (
 	"flag"
 	"fmt"
 	"log"
+	"mmd/v2mngo/db"
 	"mmd/v2mngo/tg"
 	"mmd/v2mngo/v2rpc"
 	"net/http"
+	"net/url"
 	"strconv"
-	"strings"
 
-	"github.com/redis/go-redis/v9"
 	"google.golang.org/grpc"
+	"gorm.io/gorm"
 	"v2ray.com/core/common/uuid"
 )
 
-var RDB *redis.Client
 var ctx = context.Background()
 var bt tg.Bot
 var cc *grpc.ClientConn
+var DB *gorm.DB
 
 type Userinfo struct {
 	Usedbwpretty string
@@ -42,92 +43,93 @@ func ByteCountSI(b int64) string {
 		float64(b)/float64(div), "kMGTPE"[exp])
 }
 
-func procIncome(update tg.Update, cc *grpc.ClientConn, domain string, name string) {
+func HandleUpdate(update tg.Update, cc *grpc.ClientConn, domain string, name string) {
 
 	switch update.Message.Text {
 	case "/start":
 		{
-			userUUid, err := RDB.Get(ctx, strconv.Itoa(update.Message.From.Id)).Result()
-			if err != nil {
+			user := db.User{TgId: uint(update.Message.From.Id)}
+			result := DB.FirstOrCreate(&user)
+
+			if result.RowsAffected == 1 {
 				new_uuid := uuid.New()
 				_, err := v2rpc.Adduser(new_uuid.String(), strconv.Itoa(update.Message.From.Id), cc)
 				if err != nil {
-					bt.SendMessage("failed", update.Message.From.Id)
-					log.Print("err: ", err)
+					bt.SendMessage("502 failed", update.Message.From.Id)
+					log.Printf("main: HandleUpdated: v2rpcadduser: err: %s\n", err)
 					return
 				}
-				RDB.Set(ctx, strconv.Itoa(update.Message.From.Id), new_uuid.String(), 0)
+				user.UUID = new_uuid.String()
+				DB.Save(&user)
+
+				//wellcome message
 				bt.SendMessage(fmt.Sprintf("mmdta.ir \n@naharlo \n- /start\n- /revoke\n\nhttps://%s/stat.html?uuid=%s&srv=%s", domain, new_uuid.String(), name), update.Message.From.Id)
 
 			} else {
-				if userUUid == "BlOCKED" {
-					bt.SendMessage("you are blocked", update.Message.From.Id)
+				if user.Blocked == true {
+					bt.SendMessage(fmt.Sprintf("you are blocked due to %s \n", user.LastBlockedReason), update.Message.From.Id)
 				} else {
 					used := v2rpc.GetUserStat(strconv.Itoa(update.Message.From.Id), cc)
-					bt.SendMessage(fmt.Sprintf("uuid: %s \n\ntransfered: %s", userUUid, ByteCountSI(int64(used))), update.Message.From.Id)
-
+					bt.SendMessage(fmt.Sprintf("uuid: %s \n\ntransfered: %s", user.UUID, ByteCountSI(int64(used))), update.Message.From.Id)
 				}
-
 			}
 
 		}
 	case "/revoke":
 		{
-			userUUid, _ := RDB.Get(ctx, strconv.Itoa(update.Message.From.Id)).Result()
-			if userUUid == "BLOCKED" {
-				bt.SendMessage("you are blocked", update.Message.From.Id)
+			user := db.User{TgId: uint(update.Message.From.Id)}
+			DB.First(&user)
 
+			if user.Blocked {
+				return
 			} else {
 				v2rpc.RemoveUser(strconv.Itoa(update.Message.From.Id), cc)
 				new_uuid := uuid.New()
 				_, err := v2rpc.Adduser(new_uuid.String(), strconv.Itoa(update.Message.From.Id), cc)
 				if err != nil {
-					bt.SendMessage("failed", update.Message.From.Id)
-					log.Print("err: ", err)
+					bt.SendMessage("502 failed", update.Message.From.Id)
+					log.Print("HandleUpdate: revoke: err: ", err)
 					return
 				}
-				RDB.Set(ctx, strconv.Itoa(update.Message.From.Id), new_uuid.String(), 0)
+				user.UUID = new_uuid.String()
+				DB.Save(&user)
 				bt.SendMessage(fmt.Sprintf("new uuid generated \n\nhttps://%s/stat.html?uuid=%s&srv=%s", domain, new_uuid.String(), name), update.Message.From.Id)
 			}
 		}
 
 	}
 }
+
 func startup(cc *grpc.ClientConn) {
-	iter := RDB.Scan(ctx, 0, "*", 0).Iterator()
-	for iter.Next(ctx) {
-		userUUid, err := RDB.Get(ctx, iter.Val()).Result()
-		if err != nil {
-			log.Print("err startup iter: ", err)
-		}
-		if userUUid != "BLOCKED" {
-			v2rpc.Adduser(userUUid, iter.Val(), cc)
-		}
+	var users []db.User
+	DB.Find(&users, "blocked: ?", false)
+	for _, user := range users {
+
+		v2rpc.Adduser(strconv.Itoa(int(user.TgId)), user.UUID, cc)
 	}
 
 }
 
 func main() {
+
 	tgToken := flag.String("tg", "bot123", "telegram token")
-	redisaddr := flag.String("rdis", "redis:6379", "redis addr")
 	v2raygrpc := flag.String("v2", "v2ray:8080", "v2ray grpc endpoint address")
 	name := flag.String("name", "usa", "name")
-
 	domain := flag.String("d", "tci.news", "domain")
+	database_path := flag.String("db", "db.db", "database")
+
 	flag.Parse()
 
 	bt = tg.Bot{
 		Token: *tgToken,
 	}
 
-	RDB = redis.NewClient(&redis.Options{
-		Addr:     *redisaddr,
-		Password: "",
-		DB:       0,
-	})
-
+	datab, err := db.GetDB(*database_path)
+	DB = datab
+	if err != nil {
+		log.Fatal(err)
+	}
 	cc = v2rpc.GetGrpcConn(*v2raygrpc)
-
 	startup(cc)
 
 	http.HandleFunc(fmt.Sprintf("/v2api/%s/", *tgToken), func(w http.ResponseWriter, r *http.Request) {
@@ -140,60 +142,79 @@ func main() {
 					fmt.Printf("err: %v\n", err)
 				}
 				w.WriteHeader(http.StatusOK)
-				go procIncome(up, cc, *domain, *name)
+				go HandleUpdate(up, cc, *domain, *name)
 			}
 		default:
 			{
-				urlPath := r.URL.Path
+				qs, _ := url.ParseQuery(r.URL.RawQuery)
 
-				parts := strings.Split(urlPath, "/")
-				if len(parts) != 5 {
-					var usrlis []Userinfo
-					iter := RDB.Scan(ctx, 0, "*", 0).Iterator()
-					for iter.Next(ctx) {
-						used := v2rpc.GetUserStat(iter.Val(), cc)
-						if used != 0 {
-							userid, _ := strconv.Atoi(iter.Val())
-							uo := Userinfo{Usedbwpretty: ByteCountSI(int64(used)), Usedbw: int(used), UserId: userid}
-							usrlis = append(usrlis, uo)
-						}
+				cmd, ok := qs["cmd"]
+
+				if !ok {
+					var users []db.User
+					DB.Find(&users, "blocked: ?", false)
+					for _, user := range users {
+
+						current_transfer := v2rpc.GetUserStat(strconv.Itoa(int(user.TgId)), cc)
+						user.Quoata = uint(current_transfer)
 
 					}
-					j, _ := json.Marshal(usrlis)
+					j, _ := json.Marshal(users)
 					w.Write(j)
 					return
 				}
-				cmd := parts[3]
-				uid := parts[4]
-				switch cmd {
+
+				switch cmd[0] {
 				case "block":
 					{
-						_, err := RDB.Get(ctx, uid).Result()
-						if err != nil {
-							w.Write([]byte(err.Error()))
+						tgid, ok := qs["tgid"]
+						if ok {
+							w.Write([]byte("no tgid"))
 							return
 						}
-						_, err = v2rpc.RemoveUser(uid, cc)
-						RDB.Set(ctx, uid, "BLOCKED", 0)
-						w.Write([]byte(err.Error()))
+						tgidint, _ := strconv.Atoi(tgid[0])
+						user := db.User{TgId: uint(tgidint)}
+						DB.First(&user)
+
+						_, err = v2rpc.RemoveUser(user.UUID, cc)
+						if err != nil {
+							w.Write([]byte("failed"))
+							return
+						}
+						user.Blocked = true
+						blr, ok := qs["blr"]
+						if ok {
+							user.LastBlockedReason = blr[0]
+						}
+						DB.Save(&user)
+						w.Write([]byte(fmt.Sprintf("blocked user")))
 					}
 				case "unblock":
 					{
-						uidint, err := strconv.Atoi(uid)
-						if err != nil {
-							w.Write([]byte(err.Error()))
+						tgid, ok := qs["tgid"]
+						if ok {
+							w.Write([]byte("no tgid"))
 							return
 						}
+						tgidint, _ := strconv.Atoi(tgid[0])
+						user := db.User{TgId: uint(tgidint)}
+						DB.First(&user)
+
 						new_uuid := uuid.New()
-						_, err = v2rpc.Adduser(new_uuid.String(), uid, cc)
+						_, err = v2rpc.Adduser(new_uuid.String(), tgid[0], cc)
+
 						if err != nil {
-							bt.SendMessage("failed to unblock", uidint)
-							w.Write([]byte(err.Error()))
+							//bt.SendMessage("failed to unblock", uidint)
+							w.Write([]byte(fmt.Sprintf("failed to unblock %s", err)))
 							return
 						}
-						RDB.Set(ctx, uid, new_uuid.String(), 0)
-						bt.SendMessage(fmt.Sprintf("new uuid generated \n\nhttps://%s/stat.html?uuid=%s&srv=%s", *domain, new_uuid.String(), *name), uidint)
+
+						user.UUID = new_uuid.String()
+						DB.Save(user)
+
+						bt.SendMessage(fmt.Sprintf("new uuid generated \n\nhttps://%s/stat.html?uuid=%s&srv=%s", *domain, new_uuid.String(), *name), tgidint)
 						w.Write([]byte("unblocked"))
+						return
 					}
 				}
 
